#71 Use u32 instead of usize for paging offests for search & handle search for Library

+++ packages/music_api/models/src/search/api.rs
@@ -65,7 +65,7 @@ pub enum ApiGlobalSearchResult {
 #[serde(rename_all = "camelCase")]
 #[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
 pub struct ApiSearchResultsResponse {
-    pub position: usize,
+    pub position: u32,
     pub results: Vec<ApiGlobalSearchResult>,
 }
 
@@ -81,7 +81,7 @@ impl From<Vec<ApiGlobalSearchResult>> for ApiSearchResultsResponse {
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ApiRawSearchResultsResponse {
-    pub position: usize,
+    pub position: u32,
     pub results: Vec<NamedFieldDocument>,
 }
